+++ packages/components/src/base/Accordion/accordion.tsx
@@ -1,4 +1,4 @@
 import { Accordion as MuiAccordion, type AccordionProps } from '@mui/material';
-export function Accordion(props: AccordionProps) {
+export function Accordion(props: AccordionProps): JSX.Element {
   return <MuiAccordion {...props} />;
 }