+++ packages/components/src/base/Avatar/avatargroup.tsx
@@ -1,5 +1,5 @@
 import { AvatarGroup as MuiAvatarGroup, type AvatarGroupProps } from '@mui/material';
 
-export function AvatarGroup(props: AvatarGroupProps) {
+export function AvatarGroup(props: AvatarGroupProps): JSX.Element {
   return <MuiAvatarGroup {...props} />;
 }