+++ packages/components/src/base/Avatar/badge.tsx
@@ -1,5 +1,5 @@
 import { Badge as MuiBadge, type BadgeProps } from '@mui/material';
 
-export function Badge(props: BadgeProps) {
+export function Badge(props: BadgeProps): JSX.Element {
   return <MuiBadge {...props} />;
 }