+++ packages/components/src/base/Box/box.tsx
@@ -1,5 +1,5 @@
 import { Box as MuiBox, type BoxProps } from '@mui/material';
 
-export function Box(props: BoxProps) {
+export function Box(props: BoxProps): JSX.Element {
   return <MuiBox {...props} />;
 }