+++ packages/components/src/base/Button/buttongroup.tsx
@@ -1,5 +1,5 @@
 import { ButtonGroup as MuiButtonGroup, type ButtonGroupProps } from '@mui/material';
 
-export function ButtonGroup(props: ButtonGroupProps) {
+export function ButtonGroup(props: ButtonGroupProps): JSX.Element {
   return <MuiButtonGroup {...props} />;
 }