+++ packages/components/src/base/Card/cardcontent.tsx
@@ -1,5 +1,5 @@
 import { CardContent as MuiCardContent, type CardContentProps } from '@mui/material';
 
-export function CardContent(props: CardContentProps) {
+export function CardContent(props: CardContentProps): JSX.Element {
   return <MuiCardContent {...props} />;
 }