+++ packages/components/src/base/Divider/divider.tsx
@@ -1,5 +1,5 @@
 import { Divider as MuiDivider, type DividerProps } from '@mui/material';
 
-export function Divider(props: DividerProps) {
+export function Divider(props: DividerProps): JSX.Element {
   return <MuiDivider {...props} />;
 }