+++ packages/components/src/base/ToggleButton/togglebutton.tsx
@@ -1,5 +1,5 @@
 import { ToggleButton as MuiToggleButton, type ToggleButtonProps } from '@mui/material';
 
-export function ToggleButton(props: ToggleButtonProps) {
+export function ToggleButton(props: ToggleButtonProps): JSX.Element {
   return <MuiToggleButton {...props} />;
 }