+++ packages/components/src/base/Toolbar/toolbar.tsx
@@ -1,5 +1,5 @@
 import { Toolbar as MuiToolbar, type ToolbarProps } from '@mui/material';
 
-export function Toolbar(props: ToolbarProps) {
+export function Toolbar(props: ToolbarProps): JSX.Element {
   return <MuiToolbar {...props} />;
 }