+++ packages/components/src/base/Typography/typography.tsx
@@ -1,5 +1,5 @@
 import { Typography as MuiTypography, type TypographyProps } from '@mui/material';
 
-export function Typography(props: TypographyProps) {
+export function Typography(props: TypographyProps): JSX.Element {
   return <MuiTypography {...props} />;
 }