+++ packages/components/src/custom/Dialog/index.tsx
@@ -1,0 +1,6 @@
+import StyledDialog from './StyledDialog';
+import StyledDialogActions from './StyledDialogActions';
+import StyledDialogContent from './StyledDialogContent';
+import StyledDialogTitle from './StyledDialogTitle';
+
+export { StyledDialog, StyledDialogActions, StyledDialogContent, StyledDialogTitle };