+++ packages/components/CHANGELOG.md
@@ -3,6 +3,106 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.9.0-beta.0](https://github.com/layer5io/sistent/compare/@layer5/sistent-components@0.8.1-beta.10...@layer5/sistent-components@0.9.0-beta.0) (2023-11-08)
+
+### Features
+
+- **components:** create InputLabel component ([77d51f1](https://github.com/layer5io/sistent/commit/77d51f1a6c9e8a6e99ea2dda15f76356e28f81d4)), closes [#254](https://github.com/layer5io/sistent/issues/254)
+
+## [0.8.1-beta.10](https://github.com/layer5io/sistent/compare/@layer5/sistent-components@0.8.1-beta.9...@layer5/sistent-components@0.8.1-beta.10) (2023-11-04)
+
+**Note:** Version bump only for package @layer5/sistent-components
+
+## [0.8.1-beta.9](https://github.com/layer5io/sistent/compare/@layer5/sistent-components@0.8.1-beta.8...@layer5/sistent-components@0.8.1-beta.9) (2023-11-04)
+
+### Bug Fixes
+
+- redo const to function ([056f675](https://github.com/layer5io/sistent/commit/056f675fdbd4c964ab07f4ce06b266d7dab53112))
+
+## [0.8.1-beta.8](https://github.com/layer5io/sistent/compare/@layer5/sistent-components@0.8.1-beta.7...@layer5/sistent-components@0.8.1-beta.8) (2023-11-03)
+
+### Bug Fixes
+
+- remove react-error-boundary package ([6ea9422](https://github.com/layer5io/sistent/commit/6ea9422a8125b67d1241dd6a1c323f8830a509a3))
+
+## [0.8.1-beta.7](https://github.com/layer5io/sistent/compare/@layer5/sistent-components@0.8.1-beta.6...@layer5/sistent-components@0.8.1-beta.7) (2023-11-03)
+
+### Bug Fixes
+
+- **components:** refactor components ([5b90831](https://github.com/layer5io/sistent/commit/5b90831f02644bbaa16de6e59a11a3ea50f17e58))
+
+## [0.8.1-beta.6](https://github.com/layer5io/sistent/compare/@layer5/sistent-components@0.8.1-beta.5...@layer5/sistent-components@0.8.1-beta.6) (2023-11-02)
+
+**Note:** Version bump only for package @layer5/sistent-components
+
+## [0.8.1-beta.5](https://github.com/layer5io/sistent/compare/@layer5/sistent-components@0.8.1-beta.4...@layer5/sistent-components@0.8.1-beta.5) (2023-11-02)
+
+### Bug Fixes
+
+- hardcode svg version ([#214](https://github.com/layer5io/sistent/issues/214)) ([9ea5966](https://github.com/layer5io/sistent/commit/9ea59667f9eaa50c8df846eb2477166633d0b0fb))
+
+## [0.8.1-beta.4](https://github.com/layer5io/sistent/compare/@layer5/sistent-components@0.8.1-beta.3...@layer5/sistent-components@0.8.1-beta.4) (2023-11-02)
+
+### Bug Fixes
+
+- **components:** move svg package to dep ([#213](https://github.com/layer5io/sistent/issues/213)) ([ecacb63](https://github.com/layer5io/sistent/commit/ecacb63bb65af29e650629bf0cb84f472ae6866d))
+
+## [0.8.1-beta.3](https://github.com/layer5io/sistent/compare/@layer5/sistent-components@0.8.1-beta.2...@layer5/sistent-components@0.8.1-beta.3) (2023-11-02)
+
+### Bug Fixes
+
+- **components:** support react 17 and 18 ([#211](https://github.com/layer5io/sistent/issues/211)) ([db071a1](https://github.com/layer5io/sistent/commit/db071a127d053e6da1467ef4c19f3efeb122e6a3))
+
+## [0.8.1-beta.2](https://github.com/layer5io/sistent/compare/@layer5/sistent-components@0.8.1-beta.1...@layer5/sistent-components@0.8.1-beta.2) (2023-11-02)
+
+### Bug Fixes
+
+- export mui datatables ([#209](https://github.com/layer5io/sistent/issues/209)) ([0ee51b2](https://github.com/layer5io/sistent/commit/0ee51b21fbd5a99eb923910a7fa337b1bb3129e5))
+
+## [0.8.1-beta.1](https://github.com/layer5io/sistent/compare/@layer5/sistent-components@0.8.1-beta.0...@layer5/sistent-components@0.8.1-beta.1) (2023-11-02)
+
+### Bug Fixes
+
+- **components:** hardcode svg package version ([#207](https://github.com/layer5io/sistent/issues/207)) ([11bf98b](https://github.com/layer5io/sistent/commit/11bf98bb135a1378545ee5617bfbebfa81ffcd93))
+
+## [0.8.1-beta.0](https://github.com/layer5io/sistent/compare/@layer5/sistent-components@0.8.0...@layer5/sistent-components@0.8.1-beta.0) (2023-11-02)
+
+### Bug Fixes
+
+- **components:** refactored vite config in components ([#204](https://github.com/layer5io/sistent/issues/204)) ([ee88dae](https://github.com/layer5io/sistent/commit/ee88dae5c6c897b6d47570f0ca44d3c2e4542293))
+- **svg:** redo svg pkg vite config ([#202](https://github.com/layer5io/sistent/issues/202)) ([d3ac6bb](https://github.com/layer5io/sistent/commit/d3ac6bb4904b57b4b539f1335f74b65ffef3555a))
+
+# [0.8.0](https://github.com/layer5io/sistent/compare/@layer5/sistent-components@0.6.0...@layer5/sistent-components@0.8.0) (2023-10-31)
+
+### Bug Fixes
+
+- **components:** export toolbar ([#186](https://github.com/layer5io/sistent/issues/186)) ([74c8330](https://github.com/layer5io/sistent/commit/74c83301fc9a40d7a0c36a076fb8a91c7e6ad2cc))
+- correct sx properties for select and popper ([dbd37ee](https://github.com/layer5io/sistent/commit/dbd37ee4a41905f26979022cb6b21e454e2f294f))
+
+### Features
+
+- **customcolumn:** add column visibility ([11cd979](https://github.com/layer5io/sistent/commit/11cd9794f8de9ad5fa1a3075074cf22cc1865fbf))
+- **filter:** add custom filter ([0c02847](https://github.com/layer5io/sistent/commit/0c0284794a9f46e1a45efe2a034e7201244d1505))
+- incorporate changes to ErrorBoundary ([fcbbe77](https://github.com/layer5io/sistent/commit/fcbbe77e906e31720611e0e7dfcb1fe99bcbcb7f))
+- **search-bar:** add custom search bar component ([088e92c](https://github.com/layer5io/sistent/commit/088e92c79eaf84da04de876dada98d198ad55f99))
+- **search-bar:** fix issues ([8c0c50b](https://github.com/layer5io/sistent/commit/8c0c50bec00a4da818e8521296d17a9b0e32ddb4))
+- **tooltip:** conditional tooltip ([9db1729](https://github.com/layer5io/sistent/commit/9db172934bb2437b150f8dffaca6311643d95be2))
+
+# [0.7.0](https://github.com/layer5io/sistent/compare/@layer5/sistent-components@0.6.0...@layer5/sistent-components@0.7.0) (2023-10-30)
+
+### Bug Fixes
+
+- correct sx properties for select and popper ([dbd37ee](https://github.com/layer5io/sistent/commit/dbd37ee4a41905f26979022cb6b21e454e2f294f))
+
+### Features
+
+- **customcolumn:** add column visibility ([11cd979](https://github.com/layer5io/sistent/commit/11cd9794f8de9ad5fa1a3075074cf22cc1865fbf))
+- **filter:** add custom filter ([0c02847](https://github.com/layer5io/sistent/commit/0c0284794a9f46e1a45efe2a034e7201244d1505))
+- incorporate changes to ErrorBoundary ([fcbbe77](https://github.com/layer5io/sistent/commit/fcbbe77e906e31720611e0e7dfcb1fe99bcbcb7f))
+- **search-bar:** add custom search bar component ([088e92c](https://github.com/layer5io/sistent/commit/088e92c79eaf84da04de876dada98d198ad55f99))
+- **search-bar:** fix issues ([8c0c50b](https://github.com/layer5io/sistent/commit/8c0c50bec00a4da818e8521296d17a9b0e32ddb4))
+- **tooltip:** conditional tooltip ([9db1729](https://github.com/layer5io/sistent/commit/9db172934bb2437b150f8dffaca6311643d95be2))
+
 # 0.6.0 (2023-10-16)
 
 ### Bug Fixes